requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
biopython==1.83
    # via -r requirements.in
bottleneck==1.3.8
    # via -r requirements.in
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.3
    # via -r requirements.in
commonmark==0.9.1
    # via rich
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dill==0.3.8
    # via -r requirements.in
flatdict==4.0.1
    # via -r requirements.in
fonttools==4.50.0
    # via matplotlib
gitdb==4.0.11
    # via gitpython
gitpython==3.1.42
    # via -r requirements.in
idna==3.6
    # via requests
importlib-metadata==4.13.0
    # via textual
jinja2==3.1.3
    # via -r requirements.in
joblib==1.3.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.5
    # via jinja2
matplotlib==3.8.3
    # via
    #   -r requirements.in
    #   scikit-posthocs
    #   seaborn
nanoid==2.0.0
    # via textual
natsort==7.1.1
    # via
    #   -r requirements.in
    #   pyranges
ncls==0.0.68
    # via pyranges
numpy==1.26.4
    # via
    #   -r requirements.in
    #   biopython
    #   bottleneck
    #   contourpy
    #   matplotlib
    #   ncls
    #   pandas
    #   patsy
    #   pyarrow
    #   scikit-learn
    #   scikit-posthocs
    #   scipy
    #   seaborn
    #   sorted-nearest
    #   statsmodels
packaging==24.0
    # via
    #   matplotlib
    #   statsmodels
pandas==2.2.1
    # via
    #   -r requirements.in
    #   pyranges
    #   scikit-posthocs
    #   seaborn
    #   statsmodels
    #   stringdb
patsy==0.5.6
    # via statsmodels
pillow==10.2.0
    # via matplotlib
pluggy==1.4.0
    # via -r requirements.in
polars==0.20.16
    # via -r requirements.in
pyarrow==15.0.2
    # via -r requirements.in
pycomplexheatmap==1.6.7
    # via -r requirements.in
pydantic==2.6.4
    # via -r requirements.in
pydantic-core==2.16.3
    # via pydantic
pygments==2.17.2
    # via rich
pympler==1.0.1
    # via -r requirements.in
pyparsing==3.1.2
    # via matplotlib
pyranges==0.0.129
    # via -r requirements.in
pysam==0.22.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via -r requirements.in
requests==2.31.0
    # via stringdb
rich==12.6.0
    # via
    #   -r requirements.in
    #   textual
scikit-learn==1.4.1.post1
    # via -r requirements.in
scikit-posthocs==0.9.0
    # via -r requirements.in
scipy==1.12.0
    # via
    #   -r requirements.in
    #   scikit-learn
    #   scikit-posthocs
    #   seaborn
    #   statsmodels
seaborn==0.11.2
    # via
    #   -r requirements.in
    #   scikit-posthocs
six==1.16.0
    # via
    #   patsy
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sorted-nearest==0.0.39
    # via pyranges
statsmodels==0.14.1
    # via
    #   -r requirements.in
    #   scikit-posthocs
stringdb==0.1.5
    # via -r requirements.in
tabulate==0.9.0
    # via pyranges
textual==0.4.0
    # via -r requirements.in
threadpoolctl==3.4.0
    # via scikit-learn
typing-extensions==4.10.0
    # via
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
voluptuous==0.14.2
    # via -r requirements.in
zipp==3.18.1
    # via importlib-metadata
